.github/workflows: use new ARM64 runners
In commit 0044d78d6c516c9274125f801d3c8049ff20aa01 of PR #4728 we
added the "build-docker-cross" job both our CI and release GitHub
Actions workflows in order to build Debian Linux packages of Git LFS
on the ARM64 platform.  We only build an ARM64 package for Debian 12
because the build process is very slow, as we have had to use the
Quick Emulator (QEMU) on AMD64 (x86_64) runners.

Lately, some of these CI jobs have failed due to segmentation faults
reported from the gcc compiler when it attempts to compile various
modules from the Go standard packages.  These failures appear to be
unrelated to any changes in our code, but may be due to issues with
the ARM64 emulation or some other dependency.

Fortunately, GitHub has recently made ARM64 Actions runners available
for use in public repositories like ours:

  https://github.com/orgs/community/discussions/148648

This means we can simply replace our "build-docker-cross" jobs with
"build-docker-arm" jobs that utilize the ARM64 runners, and which do
not exhibit the same kinds of failures with the gcc compiler.

The build time for these new jobs is significantly faster than the old
ones which ran in emulation, to the point where it now becomes feasible
for us to build a Debian 11 package as well as the Debian 12 one.
(We skip building a Debian 10 package since we expect to drop support
for that platform fairly soon.)

We should also be able to build an RPM package for ARM64 using Rocky 9,
once we make some updates to our RPM package metadata files and the
corresponding Dockerfile in our git-lfs/build-dockers repository.
For the time being, however, we defer these changes to future PRs.