commands/command_fetch.go: reorder watch channel
In prior commits in this PR we refactored and simplified the
fetchAndReportToChan() function in our "commands" package so it is
now just called fetch(), and also revised this function to introduce
support for a new --dry-run option to our "git lfs fetch" command.

When the --dry-run command-line option is specified, the fetch() function
starts an anonymous goroutine to output the path and object ID of each
Git LFS object the command would normally retrieve via a transfer queue.

At the moment, this goroutine calls the Watch() method of the TransferQueue
structure in our "tq" package to create and register a "watcher" channel
with the queue, and then starts a loop to read the values sent over the
channel by the queue, each of which represents an object the queue
would transfer if it was not executing in a dry-run mode.

So long as the goroutine calls the Watch() method before the transfer
queue begins operating, this design works as we intend.  However, if
the goroutine's execution is delayed by the Go scheduler, the queue
may perform all of its actions before the goroutine calls the Watch()
method, and the fetch() function could invoke the Wait() method of
the TransferQueue structure before the goroutine calls the Watch()
method.

Should this happen, the "watcher" channel created by the Watch()
method will never be closed, because that only happens in the Wait()
method, and that method has already run at a time when there were
no extant channels to close.  Therefore the goroutine's loop over the
channel will never exit, and the goroutine will never decrement the
count of the WaitGroup structure by calling its Done() method, and in
turn the fetch() function's call to the WaitGroup's Wait() method will
never complete, so the "git lfs fetch" command will stall indefinitely.

To ensure we handle this scheduling condition, we create the "watcher"
channel in the fetch() function before starting our anonymous goroutine,
so the channel will always be registered with the TransferQueue before
its Wait() method is called.  Thus the channel will always be found
and closed by that method, and the goroutine cannot enter an infinite
wait state on the channel.