.github/workflows/ci.yml: find prove on Windows
As described in commit 8d456a76727bbadf690b0093e1343331bff9ac23 of
PR #5879, we currently depend on the Git for Windows SDK to install a
version of the "prove" command from the Perl Test::Harness distribution,
because we use that command to run our suite of shell tests.

However, in our CI jobs on Windows, the SDK does not at present install
the "prove" command into a location listed in the PATH environment variable,
so these CI jobs having begun failing.

This change may be related to the recent migration of the SDK's
release artifacts to GitHub from Azure Blobs, as outlined in
PR git-for-windows/git-for-windows-automation#109, and is possibly
a consequence of PR git-for-windows/build-extra#588.

Regardless, the "prove" command is fortunately still included in the
"minimal" flavour of the SDK, since it is listed in the manifest for
that flavour:

  https://github.com/git-for-windows/git-sdk-64/blob/f845bb725e56e058a0fbf93aba18c70906a1068e/.sparse/minimal-sdk#L188

We simply need to ensure that the directory in which the command is
now located, namely /usr/bin/core_perl, is added to the set of paths
in the PATH variable for the Git Bash environment.

Note that we only need to add this extra path to PATH when we are
running our script/cibuild script, as that runs the default recipe
from our t/Makefile file, which in turn executes the "test" recipe,
which uses the "prove" command to run and summarize all our shell
test scripts.