.github/workflows: use Ruby provided by runners
In PR #3840 we introduced a GitHub Actions workflow to build our
release assets when we push a new version tag.  As a final step in this
workflow we run our script/packagecloud.rb script, which requires the
use of the "packagecloud-ruby" Ruby gem.  At the time, these
requirements meant we needed to first execute the actions/setup-ruby
action so we could then install the gem and run our Ruby script.

Later, in PR #4230, we added steps to our CI workflow to build our
manual pages, and since these required the use of the "ronn" Ruby gem,
we also added the actions/setup-ruby action to our CI workflow.
(We subsequently migrated our manual page source files to the AsciiDoc
format, in PR #5054, but we continue to use Ruby and the "asciidoctor"
gem to build our manual pages.)

Then in commit b7fa3a54e4d65e5f000b30e9df41cddf1daeed41 of PR #5236 we
upgraded both of our Actions workflows to use the ruby/setup-ruby
action instead of the now-deprecated actions/setup-ruby one.  Because
the ruby/setup-ruby action installs the MSYS2 environment on Windows
and sets several key environment variables like PATH and TMPDIR, we
also introduced steps to make sure our CI and release processes
continued to work as expected in this context, by clearing the
TMPDIR variable and renaming the directory containing the MSYS2
executables.

Fortunately, the default runners provided by GitHub Actions for the
macOS, Windows, and Ubuntu Linux platforms are now all provisioned with
an installation of Ruby 3.x which includes the "gem" utility we need to
install the "asciidoctor" and "packagecloud-ruby" gems.  As these are
all we require to run our Ruby scripts, we no longer need the more
extensive Ruby development environment provided by the ruby/setup-ruby
action.

We can therefore simply remove the ruby/setup-ruby steps from our
workflows, along with all the special handling of the MSYS2 environment
and the TMPDIR environment variable on Windows.