docs/howto/release-git-lfs.md: adjust formatting
In the documentation of our release processes, we describe our release
types using the capitalized terms MAJOR, MINOR, and PATCH; however, we
do not capitalize the term PATCH consistently, so we adjust that now.

As well, we also add language identifiers to several Markdown code
blocks with currently lack them.