docs/howto/release-git-lfs.md: add security steps
The documentation of our standard and patch release processes does not,
at present, include the steps required to build a security patch release,
so we add two sections which describe how we build such releases.

The first new section documents our normal security patch release
process, which applies when we do not need to coordinate our release
with other projects, or at least do not need to build release assets
in private while under a publication embargo.

The second new section details the process of building new Git LFS
release binaries in private, so they can be shared with other downstream
projects while we are under an embargo on the publication of the
security vulnerability and associated security patch release.