commands: replace unused debug mode with tracing
In PR #68 the "commands" package was updated with several logging
functions such as Panic(), Print(), and Debug().  The last of these
only outputs a log message if the Debugging variable in the same
package is set to "true", which at the time was controlled by the
setupDebugging() function.  If the user supplied the --debug option
to a given Git LFS command, setupDebugging() function would set the
Debugging variable to "true", and any calls to Debug() would then
generate log messages.

However, the setupDebugging() function was removed, along with a
lot of the earlier command-line parsing and setup code, in PR #84,
and replaced with the Cobra library from the spf13/cobra project.
Since then it has not been possible to set the Debugging variable
to a value other than "false", so the remaining references to the
Debug() function will never generate log messages.

We therefore remove the Debugging variable and the Debug() function,
and replace any calls to that function with ones to our standard
trace logging function.