avoid new error format strings without arguments
When we report an error, we often call the Errorf() function in our
custom "errors" package, which accepts both a format string and a
variable number of additional arguments.  The intent of this function
is for the additional arguments to be interpolated into the format
string, and then an error created using the string which results from
the interpolation.

In many instances we originally made use of the interpolation capability
of the Errorf() function by passing at least one additional argument to
it other than the format string.  In other cases, though, we used the
Errorf() function with just a single fixed string argument, so the
the New() function of the same "errors" package would have sufficed
instead.

In PR #4781 we then updated our code to allow for localization of most
of our output messages by inserting calls to the Get() method of the Tr
structure of our "tr" package.  In the instances where we call the
Errorf() function, the consequence is that we now pass the format string
and any additional arguments to the Get() method, and then just pass
the string it returns to the Errorf() function as a single argument.

As of Go 1.24, if the Go version number in the "go.mod" file is set to
1.24 or higher, the "go vet" command now reports misuses of non-constant
strings as format strings.  This includes all the cases where we pass
the output from the Get() method of the Tr structure to an Errorf()
function as a single argument.  To resolve this concern, we revise
all such instances to use the New() function of our "errors" package
in place of the Errorf() function.