tasklog/simple_task.go: supply format string
As of Go 1.24, if the Go version number in the "go.mod" file is set to
1.24 or higher, the "go vet" command now reports misuses of non-constant
strings as format strings.

In the Log() method of the SimpleTask structure in our "tasklog" package,
we pass the method's non-constant input string as the only argument to
the Logf() method, which then uses it as a format string when creating
a task log Update structure.

To resolve this issue, we instead call the Logf() method with a format
string containing a single "%s" verb, and pass the non-constant input
string as the second argument.