t/git-lfs-test-server-api: use format string
As of Go 1.24, if the Go version number in the "go.mod" file is set to
1.24 or higher, the "go vet" command now reports misuses of non-constant
strings as format strings.

Our Git LFS Batch API test utility, the git-lfs-test-server-api program,
defines an exit() function which accepts a format string and a variable
number of arguments, generates an output message from those, and then
halts the program with a non-zero exit code.  In one instance we call
this exit() function with a non-constant format string we construct
from a fixed string and an error message.

To resolve this issue, we revise the fixed string to include a "%s"
verb, and pass the error as a second argument to the exit() function.