commands/command_pointer.go: avoid format string
As of Go 1.24, if the Go version number in the "go.mod" file is set to
1.24 or higher, the "go vet" command now reports misuses of non-constant
strings as format strings.

When the "git lfs pointer" command compares two Git LFS pointer files,
it outputs the contents of the file specified by the --pointer option
(or supplied on standard input, if the --stdin option is used) to the
standard error file descriptor.  To write this data, we call the
Fprintf() function of the Go standard library's "fmt" package, and
pass the pointer file data as the format string, although it is not
a constant string value.

To resolve this issue, we simply need to use the Fprint() function
instead, as we already do in a number of other instances in the same
command where we write data to the standard error file descriptor.