git/githistory: avoid exporting refUpdater fields
The refUpdater structure and related methods were introduced to the
"githistory" package in PR #2340, as part of the series of PRs that
implemented the "git lfs migrate" command.  The structure is not
exported outside of the package, as it is only used by the Rewrite()
method of the Rewriter structure in the same package.  (The Rewriter
structure and methods are the principal external interface of the
"githistory" package.)

While the structure itself is not exported, most of its elements are
exported, as is its UpdateRefs() method.  However, none of these are
referenced outside of the "githistory" package, so we rename them
now to clarify that they are only used within the package.