git/githistory/rewriter.go: rename logger var
In the "githistory" package, the Rewriter and refUpdater structures both
contain elements that are pointers to Logger structures from the "tasklog"
package.  Moreover, the refUpdater structure's element of this type
is initialized from the value of the corresponding element of the
Rewriter structure.

However, these elements have different names, so we rename the Logger
element of the Rewriter structure of "logger", which brings the two
structure's element names into closer alignment, and also avoids the use
of the abbreviated name "l", since the Logger element is used in multiple
disparate methods and a more descriptive name is helpful in such cases.