commands,t: drop prune progress prefix
In PR #2757 the output of the "git lfs prune" command was revised to
use the methods of the Logger structure of the "tasklog" package to
output progress messages, replacing use of the Spinner structure from
the older "progress" package.  The Spinner structure and its methods
were then removed in PR #2759, and the "progress" package was removed
entirely in PR #2762.

As part of this change, the "git lfs prune" command's progress messages
were reformatted so as to always begin with the Git LFS subcommand name,
for example, "prune: 8 local objects, 9 retained, done."

Two other Git LFS commands follow the same design, the "git lfs fetch"
and "git lfs migrate" commands, but no others.  Git commands also do
not prefix their progress messages with the name of the subcommand
(although messages from a Git remote may be reported with the prefix
"remote:").

To help bring all our commands' progress messages into alignment, we
simply remove the "prune:" prefix from the messages output by the
"git lfs prune" command.  We expect to revise the progress messages
from the "git lfs fetch" and "git lfs migrate" commands in the same
way in subsequent commits of this PR.