commands,docs,git,t: drop migrate progress prefix
In PR #2329 the output of the "git lfs migrate" command was revised to
use the methods of the Logger structure of the "log" package to output
progress messages.  This package was originally located within the
"git/githistory" source directory, but was later moved and renamed to
the "tlog" package and then to the current "tasklog" package in PR #2747.

As part of the changes in PR #2329, the methods of the Rewriter
structure in the "githistory" package were updated to use the methods
of the Logger and PercentageTask types in the "log" package to generate
progress messages which began with the Git LFS subcommand name, for
example, "migrate: Rewriting commits".

Since that PR, multiple other informational and error messages output
by the "git lfs migrate" command have been added, and while most begin
with the "migrate:" prefix, not all do.  For instance, running the
"git lfs migrate import --verbose" command outputs commit SHAs and
file paths with the prefix, but Git references and the mapping of their
old and new SHAs without the prefix.

Two other Git LFS commands, the "git lfs prune" and "git lfs fetch"
commands, follow the same design and output their subcommand names as
prefixes to their progress messages, but no other commands do this.
Git commands also do not prefix their progress messages with the name
of the subcommand (although messages from a Git remote may be reported
with the prefix "remote:").

To help bring all our commands' progress messages into alignment, we
simply remove the "migrate:" prefix from the messages output by the
"git lfs migrate" command.  We have similarly altered the progress
messages of the "git lfs prune" and "git lfs fetch" commands in
previous commits in this PR.