commands,t: tidy verbose migrate output format
When the --verbose option is supplied to the "git lfs migrate export"
or "git lfs migrate import" command, additional output messages are
generated listing the file paths of all migrated files and the original
commit SHAs which introduced those files.  This option was first
introduced for the "git lfs migrate import" command in PR #2610.

The rewriteBlob() method of the Rewriter structure in the "githistory"
package outputs the pairs of original commits SHAs and file paths on
individual lines by adding entries to a PercentageTask structure from
the "tasklog" package.  Each line is formatted without any leading
whitespace.

During a subsequent step in the commands' progress, the updateOneRef()
method of the refUpdater structure in the "githistory" package outputs
pairs of Git reference names and the original and rewritten commit SHAs
associated with them.  Unlike the list of commits and file paths, these
lines are formatted with an indentation of two space characters, which
sets them apart visually from the main progress messages output by
the command.

To help clarify the principal steps of the migration process and
make the output of the "git lfs migrate" subcommands more readable,
we add indentation to the lines with the pairs of commit SHAs and
file paths, matching that of the lines with the Git references and
old and new commit SHAs.

We then update the "migrate export (--verbose)" test in our
t/t-migrate-export.sh test script to validate the format of the
lines with commit SHAs output by the "git lfs migrate export" command.
Similarly, we update two tests in our t/t-migrate-import.sh test script
to do the same for the "git lfs migrate import" command.

Note that in all of these tests we use the tee(1) command to capture
the output of the "git lfs migrate" command under test into a log file,
and then use a grep(1) command to check for the formatted lines in the
log file.  We avoid piping the output of the "git lfs migrate" command
directly into the grep(1) command because we prefer to use that command's
-q option, which causes it to exit as soon as it reads a matching line.
This in turn would cause the "git lfs migrate" command to exit
immediately when it receives a SIGPIPE signal trying to write to the
pipe, which would leave the migration incomplete.

Further, we need to pass the --yes option to the "git lfs migrate"
command in these tests because it will now otherwise pause and wait
for confirmation to proceed when it detects the untracked "migrate.log"
file being written into the current Git working tree.

Finally, we also update the final progress message output by the
"git lfs migrate export" and "git lfs migrate import" subcommands to
begin with an uppercase letter.  It now reads "Checkout" rather than
"checkout", which better aligns with the format of all the preceding
messages such as "Sorting commits" and "Rewriting commits".