lfs/gitfilter_smudge.go: make error translatable
In commit f97206432dd3dffae3c9e3e129bc3cecea11442f of PR #5066 we
introduced the "lfs.remote.searchAll" configuration option and updated
the "smudge" Git LFS filter logic so that when the option is enabled,
the filter searches all Git remotes for Git LFS objects that are not
present locally rather than only attempting to fetch the objects from
the default Git LFS remote URL.

The downloadFileFallBack() method of the GitFilter structure in the
"lfs" package was added to implement this logic.  One of the error
conditions it must handle may occur if no Git remotes are defined.
In this case, we append (i.e., wrap) an initial error message with
a second generic error message indicating which file and Git LFS
object ID could not be downloaded.

While we allow for the generic error message to be localized, we
do not do so for the initial error message, so we adjust that now.
We simply call the Get() method of the Tr structure in our "tr" package
on the initial error message, as we already do for the generic
error message.