avoid wrap error format strings without arguments
When we report an error, we often call the Wrapf() function in our
custom "errors" package, which accepts an existing error, a format
string, and a variable number of additional arguments.  The intent of
this function is for the additional arguments to be interpolated into
the format string, and then a new error created using the string which
results from the interpolation, combined with the initial error and
its message.

In many instances we originally made use of the interpolation capability
of the Wrapf() function by passing at least one additional argument to
it other than the existing error and the format string.  In other cases,
though, we used the Wrapf() function with just the existing error and
a fixed string argument, so the Wrap() function of the same "errors"
package would have sufficed instead.

In PR #4781 updated our code to allow for localization of most of our
output messages by inserting calls to the Get() method of the Tr
structure of our "tr" package.  In the instances where we call the
Wrapf() function, the consequence is that we now pass the format string
and any additional arguments to the Get() method, and then just pass
the string it returns to the Wrapf() function as the format string
argument following the initial existing error argument.

As of Go 1.24, if the Go version number in the "go.mod" file is set to
1.24 or higher, the "go vet" command now reports misuses of non-constant
strings as format strings.  This includes all the cases where we pass
the output from the Get() method of the Tr structure to an Wrapf()
function as a format string without any additional arguments to be
interpolated.  To resolve this concern, we revise all such instances to
use the Wrap() function of our "errors" package in place of the Wrapf()
function.