errors: drop unused StackTrace() function
In PR #1466 we introduced our custom "errors" package as a replacement
for the "errutil" package used previously to provide various error
handling functions and specialized error types.  As part of this PR,
in commit 8624a873286173a763c507f3323e679dd0bd26e6 we defined the
StackTrace() function in the new "errors" package, and updated the
logPanicToWriter() function in our "commands" package to call it
when generating an error log file.  We write these log files to
the .git/lfs/logs directory whenever a Git LFS command invokes one
of our error reporting utility functions such as LoggedError() or
Panic().

Later, in commit 377366d4953bb7be2222c12cfc620fa3caeaabe7 of PR #2178,
we revised the logPanicToWriter() function to use the "%+v" format
string specifier when writing an error's details to the log file,
and dropped the call to the StackTrace() function of our "errors"
package.

The StackTrace() call was deemed no longer necessary because we
expect all of our errors to be created using our "errors" package's
functions, which instantiate error structures defined by the legacy
"github.com/pkg/errors" package.  This package defines its own
Format() methods for its structures so that when the "%+v" format
string specifier is used with them, they output both the message
and stack trace associated with each error.

As a consequence of this change in PR #2178, the StackTrace()
function in our own custom "errors" package has no remaining callers,
so we can remove it now.